+++ src/index.ts
@@ -30,5 +30,7 @@ export { PaymentsResource } from "./resources/payments";
 export { MeasuresResource } from "./resources/measures";
 export { SkinResource } from "./resources/skin";
 export { MealsResource } from "./resources/meals";
+export { LaboratoryResource } from "./resources/laboratory";
+export { DietsResource } from "./resources/diets";
 
 export * from "./models";